add prototype of sstable

+++ src/main/java/ru/golchin/Util.java
@@ -5,11 +5,15 @@ import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import static java.nio.file.Files.exists;
+
 public final class Util {
     private Util() {
     }
 
     public static void deleteDirectory(Path p) throws IOException {
+        if (!exists(p))
+            return;
         Files.list(p).forEach(path -> {
             try {
                 Files.delete(path);
